Create single file for each table in holidays page
